refactor(subway) 소스코드 코드 컨벤션 적용
