+++ src/main/java/nextstep/subway/auth/infrastructure/AuthenticationPrincipalConfig.java
@@ -1,12 +1,11 @@
 package nextstep.subway.auth.infrastructure;
 
+import java.util.List;
 import nextstep.subway.auth.application.AuthService;
 import nextstep.subway.auth.ui.AuthenticationPrincipalArgumentResolver;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
 
 @Configuration
 public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
