+++ src/main/java/nextstep/subway/auth/infrastructure/AuthorizationExtractor.java
@@ -1,7 +1,7 @@
 package nextstep.subway.auth.infrastructure;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.Enumeration;
+import javax.servlet.http.HttpServletRequest;
 
 public class AuthorizationExtractor {
     public static final String AUTHORIZATION = "Authorization";
