+++ src/main/java/nextstep/subway/auth/ui/AuthenticationPrincipalArgumentResolver.java
@@ -1,5 +1,7 @@
 package nextstep.subway.auth.ui;
 
+import java.util.Objects;
+import javax.servlet.http.HttpServletRequest;
 import nextstep.subway.auth.application.AuthService;
 import nextstep.subway.auth.domain.AuthenticationPrincipal;
 import nextstep.subway.auth.domain.LoginMember;
@@ -10,10 +12,8 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
-import javax.servlet.http.HttpServletRequest;
-
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
-    private AuthService authService;
+    private final AuthService authService;
 
     public AuthenticationPrincipalArgumentResolver(AuthService authService) {
         this.authService = authService;
@@ -25,10 +25,12 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
     }
 
     @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
+                                  NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         AuthenticationPrincipal parameterAnnotation = parameter.getParameterAnnotation(AuthenticationPrincipal.class);
-        String credentials = AuthorizationExtractor.extract(webRequest.getNativeRequest(HttpServletRequest.class));
-        if(!parameterAnnotation.required() && credentials == null) {
+        String credentials = AuthorizationExtractor.extract(
+                Objects.requireNonNull(webRequest.getNativeRequest(HttpServletRequest.class)));
+        if (!Objects.requireNonNull(parameterAnnotation).required() && credentials == null) {
             return new LoginMember();
         }
         return authService.findMemberByToken(credentials);
