+++ src/main/java/nextstep/subway/path/application/PathFinder.java
@@ -2,7 +2,7 @@ package nextstep.subway.path.application;
 
 import java.util.Set;
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.Path;
+import nextstep.subway.path.domain.Path;
 import nextstep.subway.station.domain.Station;
 
 public interface PathFinder {
