+++ src/test/java/nextstep/subway/DomainFixtureFactory.java
@@ -6,7 +6,7 @@ import nextstep.subway.fare.domain.Fare;
 import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.line.domain.Distance;
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.Path;
+import nextstep.subway.path.domain.Path;
 import nextstep.subway.line.domain.Section;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.SectionRequest;
