+++ src/test/java/nextstep/subway/path/domain/PathTest.java
@@ -11,7 +11,6 @@ import com.google.common.collect.Lists;
 import java.util.List;
 import nextstep.subway.line.domain.Distance;
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.Path;
 import nextstep.subway.line.domain.Section;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
